requirements.txt
faiss-cpu
openai==0.28
python-dotenv
docx
SpeechRecognition
pyttsx3
pyaudio